refactor: 💡 dirs
