+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,1 +1,1 @@
-export { GET, POST } from "@/features/auth/lib/auth";
+export { GET, POST } from "@/features/auth/utils/auth";
