+++ src/app/contents/books/page.tsx
@@ -1,10 +1,10 @@
+import { getAllImages, getAllSlugs } from "@/apis/markdown/fetch-contents";
 import { Header } from "@/components/nav/header";
 import { Badge } from "@/components/ui/badge";
 import { Unauthorized } from "@/components/unauthorized";
 import { PAGE_NAME } from "@/constants";
-import { checkAdminPermission } from "@/features/auth/lib/role";
+import { checkAdminPermission } from "@/features/auth/utils/role";
 import { ContentsStack } from "@/features/contents/components/contents-stack";
-import { getAllImages, getAllSlugs } from "@/features/contents/utils/api";
 import { formatSlugsAndImages } from "@/features/contents/utils/format";
 import type { Metadata } from "next";
