+++ src/app/contents/notes/page.tsx
@@ -1,9 +1,9 @@
+import { getAllImages, getAllSlugs } from "@/apis/markdown/fetch-contents";
 import { Header } from "@/components/nav/header";
 import { Unauthorized } from "@/components/unauthorized";
 import { PAGE_NAME } from "@/constants";
-import { checkAdminPermission } from "@/features/auth/lib/role";
+import { checkAdminPermission } from "@/features/auth/utils/role";
 import { ContentsStack } from "@/features/contents/components/contents-stack";
-import { getAllImages, getAllSlugs } from "@/features/contents/utils/api";
 import { formatSlugsAndImages } from "@/features/contents/utils/format";
 import type { Metadata } from "next";
