+++ src/features/auth/actions/sign-out.tsx
@@ -3,7 +3,7 @@ import "server-only";
 import { sendLineNotifyMessage } from "@/apis/line-notify/send-message";
 import { SUCCESS_MESSAGES } from "@/constants";
 import { formatErrorForClient } from "@/error";
-import { signOut as NextAuthSignOut } from "@/features/auth/lib/auth";
+import { signOut as NextAuthSignOut } from "@/features/auth/utils/auth";
 import type { ServerAction } from "@/types";
 
 type SignOutState = ServerAction<undefined>;
