+++ src/features/contents/types.ts
@@ -1,0 +1,4 @@
+export type ContentsType = {
+	slug: string;
+	content: string;
+};
