--- src/features/contents/types/contents.ts
@@ -1,7 +1,0 @@
-// TODO: common typesディレクトリに移動+contentsをbookとnotesに命名変更
-type ContentsType = {
-	slug: string;
-	content: string;
-};
-
-export default ContentsType;
